Add files via upload
Hello guys, here I come again with another map for you! :laughing:

This map consists of an interior in the Chinatown area, to create more Roleplay in this dead area of ​​the map.

I also wanted to tell you that I am working and developing more maps. If you want to be more informed about the latter, you can join my development discord. Not only will there be maps but also Scripts for FiveM as well.
 
All credits for: Lete

**Personal Discord:** !Lete#6937

**Discord Comunity:** https://discord.gg/9hn6n4j

**Preview:** https://youtu.be/1qp1mYtFcy4